requirements.txt
altgraph==0.17.4
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
python-dateutil==2.9.0.post0
python-whois==0.9.4
requests==2.32.3
six==1.16.0
urllib3==2.2.2
Werkzeug==3.0.3
